requirements.txt
torch==1.2.0
torchvision==0.4.0
numpy==1.16.5
matplotlib==3.0.3
Pillow==4.3.0
